LayoutTests/inspector/debugger/xhr-breakpoints.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script>

function sendRequest(url)
{
    var request = new XMLHttpRequest();
    request.open("GET", url, true);
    request.send()
}

function test()
{
    var pane = WebInspector.panels.scripts.sidebarPanes.xhrBreakpoints;
    InspectorTest.runDebuggerTestSuite([
        function testXHRBreakpoint(next)
        {
            pane._setBreakpoint("foo", true);
            InspectorTest.waitUntilPaused(step1);
            InspectorTest.evaluateInPageWithTimeout("sendRequest('/foo?a=b')");

            function step1(callFrames)
            {
                InspectorTest.captureStackTrace(callFrames);
                InspectorTest.resumeExecution(step2);
            }

            function step2()
            {
                InspectorTest.evaluateInPage("sendRequest('/bar?a=b')", step3);
            }

            function step3()
            {
                pane._removeBreakpoint("foo");
                InspectorTest.evaluateInPage("sendRequest('/foo?a=b')", next);
            }
        },

        function testPauseOnAnyXHR(next)
        {
            pane._setBreakpoint("");
            InspectorTest.evaluateInPage("sendRequest('/foo?a=b')", pausedFoo);

            function pausedFoo(callFrames)
            {
                function resumed()
                {
                    InspectorTest.evaluateInPage("sendRequest('/bar?a=b')", pausedBar);
                }
                InspectorTest.resumeExecution(resumed);
            }

            function pausedBar(callFrames)
            {
                function resumed()
                {
                    pane._removeBreakpoint("");
                    InspectorTest.evaluateInPage("sendRequest('/baz?a=b')", next);
                }
                InspectorTest.resumeExecution(resumed);
            }
        }
    ]);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests XHR breakpoints.
</p>

</body>
</html>
